move review_pr function to framework.easyconfig.tools module
